# Speed up SRTP driver tests